+++ flowman-spark-extensions/src/main/scala/com/dimajix/spark/sql/RowParser.scala
@@ -184,9 +184,13 @@ class RowParser(schema: StructType, options:RowParser.Options) {
     }
 
     def parse(tokens: Array[String]): Row = {
-        parse(tokens, "")
+        parseInternal(tokens, None)
     }
     def parse(tokens: Array[String], line:String): Row = {
+        parseInternal(tokens, Some(line))
+    }
+
+    private def parseInternal(tokens: Array[String], line:Option[String]): Row = {
         // If the number of tokens doesn't match the schema, we should treat it as a malformed record.
         // However, we still have chance to parse some of the tokens, by adding extra null tokens in
         // the tail if the number is smaller, or by dropping extra tokens if the number is larger.
@@ -196,7 +200,7 @@ class RowParser(schema: StructType, options:RowParser.Options) {
             throw BadRecordException(
                 () => new UTF8String(),
                 () => None,
-                new RuntimeException(s"Malformed record. Expected ${schema.length} columns, but only got ${tokens.length} columns:\n$line"))
+                new RuntimeException(s"Malformed record. Expected ${schema.length} columns, but only got ${tokens.length} columns. Offending record:\n${line.getOrElse(tokens.mkString(","))}"))
         }
 
         val checkedTokens = if (schema.length > tokens.length) {